move all tree implementation in one file
