+++ hummingbird/ml/operator_converters/_gbdt_commons.py
@@ -13,126 +13,7 @@ import numpy as np
 
 from . import constants
 from ._tree_commons import get_tree_params_and_type, get_parameters_for_tree_trav_common, get_parameters_for_gemm_common
-from ._tree_implementations import GEMMTreeImpl, TreeTraversalTreeImpl, PerfectTreeTraversalTreeImpl, TreeImpl
-
-
-class GEMMGBDTImpl(GEMMTreeImpl):
-    """
-    Class implementing the GEMM strategy (in PyTorch) for GBDT models.
-    """
-
-    def __init__(self, net_parameters, n_features, classes=None, extra_config={}):
-        """
-        Args:
-            net_parameters: The parameters defining the tree structure
-            n_features: The number of features input to the model
-            classes: The classes used for classification. None if implementing a regression model
-            extra_config: Extra configuration used to properly implement the source tree
-        """
-        super(GEMMGBDTImpl, self).__init__(net_parameters, n_features, classes, 1)
-        self.n_gbdt_classes = 1
-
-        if constants.LEARNING_RATE in extra_config:
-            self.learning_rate = extra_config[constants.LEARNING_RATE]
-        if constants.ALPHA in extra_config:
-            self.alpha = torch.nn.Parameter(torch.FloatTensor(extra_config[constants.ALPHA]), requires_grad=False)
-
-        if classes is not None:
-            self.n_gbdt_classes = len(classes) if len(classes) > 2 else 1
-            if self.n_gbdt_classes == 1:
-                self.binary_classification = True
-
-        self.n_trees_per_class = len(net_parameters) // self.n_gbdt_classes
-
-    def aggregation(self, x):
-        return torch.squeeze(x).t().view(-1, self.n_gbdt_classes, self.n_trees_per_class).sum(2)
-
-    def calibration(self, x):
-        if self.binary_classification:
-            output = torch.sigmoid(x)
-            return torch.cat([1 - output, output], dim=1)
-        else:
-            return torch.softmax(x, dim=1)
-
-
-class TreeTraversalGBDTImpl(TreeTraversalTreeImpl):
-    """
-    Class implementing the Tree Traversal strategy in PyTorch.
-    """
-
-    def __init__(self, net_parameters, max_detph, n_features, classes=None, extra_config={}):
-        """
-        Args:
-            net_parameters: The parameters defining the tree structure
-            max_depth: The maximum tree-depth in the model
-            n_features: The number of features input to the model
-            classes: The classes used for classification. None if implementing a regression model
-            extra_config: Extra configuration used to properly implement the source tree
-        """
-        super(TreeTraversalGBDTImpl, self).__init__(net_parameters, max_detph, n_features, classes, 1)
-        self.n_gbdt_classes = 1
-
-        if constants.LEARNING_RATE in extra_config:
-            self.learning_rate = extra_config[constants.LEARNING_RATE]
-        if constants.ALPHA in extra_config:
-            self.alpha = torch.nn.Parameter(torch.FloatTensor(extra_config[constants.ALPHA]), requires_grad=False)
-
-        if classes is not None:
-            self.n_gbdt_classes = len(classes) if len(classes) > 2 else 1
-            if self.n_gbdt_classes == 1:
-                self.binary_classification = True
-
-        self.n_trees_per_class = len(net_parameters) // self.n_gbdt_classes
-
-    def aggregation(self, x):
-        return x.view(-1, self.n_gbdt_classes, self.n_trees_per_class).sum(2)
-
-    def calibration(self, x):
-        if self.binary_classification:
-            output = torch.sigmoid(x)
-            return torch.cat([1 - output, output], dim=1)
-        else:
-            return torch.softmax(x, dim=1)
-
-
-class PerfectTreeTraversalGBDTImpl(PerfectTreeTraversalTreeImpl):
-    """
-    Class implementing the Perfect Tree Traversal strategy in PyTorch.
-    """
-
-    def __init__(self, net_parameters, max_depth, n_features, classes=None, extra_config={}):
-        """
-        Args:
-            net_parameters: The parameters defining the tree structure
-            max_depth: The maximum tree-depth in the model
-            n_features: The number of features input to the model
-            classes: The classes used for classification. None if implementing a regression model
-            extra_config: Extra configuration used to properly implement the source tree
-        """
-        super(PerfectTreeTraversalGBDTImpl, self).__init__(net_parameters, max_depth, n_features, classes, 1)
-        self.n_gbdt_classes = 1
-
-        if constants.LEARNING_RATE in extra_config:
-            self.learning_rate = extra_config[constants.LEARNING_RATE]
-        if constants.ALPHA in extra_config:
-            self.alpha = torch.nn.Parameter(torch.FloatTensor(extra_config[constants.ALPHA]), requires_grad=False)
-
-        if classes is not None:
-            self.n_gbdt_classes = len(classes) if len(classes) > 2 else 1
-            if self.n_gbdt_classes == 1:
-                self.binary_classification = True
-
-        self.n_trees_per_class = len(net_parameters) // self.n_gbdt_classes
-
-    def aggregation(self, x):
-        return x.view(-1, self.n_gbdt_classes, self.n_trees_per_class).sum(2)
-
-    def calibration(self, x):
-        if self.binary_classification:
-            output = torch.sigmoid(x)
-            return torch.cat([1 - output, output], dim=1)
-        else:
-            return torch.softmax(x, dim=1)
+from ._tree_implementations import GEMMGBDTImpl, TreeTraversalGBDTImpl, PerfectTreeTraversalGBDTImpl, TreeImpl
 
 
 def convert_gbdt_classifier_common(tree_infos, get_tree_parameters, n_features, n_classes, classes=None, extra_config={}):
